+++ Engine/Renderer/Vulkan/VulkanBackend.hpp
@@ -73,11 +73,11 @@ public:
 	virtual bool CreateShader(Shader* shader, const ShaderConfig* config, IRenderpass* pass, const std::vector<char*>& stage_filenames, std::vector<ShaderStage>& stages) override;
 	virtual bool UseShader(Shader* shader) override;
 	virtual bool BindGlobalsShader(Shader* shader) override;
-	virtual bool BindInstanceShader(Shader* shader, uint32_t instance_id) override;
+	virtual bool BindInstanceShader(Shader* shader, uint64_t instance_id) override;
 	virtual bool ApplyGlobalShader(Shader* shader) override;
 	virtual bool ApplyInstanceShader(Shader* shader, bool need_update) override;
 	virtual uint32_t AcquireInstanceResource(Shader* shader, std::vector<TextureMap*>& maps) override;
-	virtual bool ReleaseInstanceResource(Shader* shader, uint32_t instance_id) override;
+	virtual bool ReleaseInstanceResource(Shader* shader, uint64_t instance_id) override;
 	virtual bool SetUniform(Shader* shader, ShaderUniform* uniform, const void* value) override;
 
 	virtual bool AcquireTextureMap(TextureMap* map) override;
