+++ Engine/Systems/ShaderSystem.h
@@ -156,7 +156,7 @@ public:
 	 * @param instance_id The identifier of the instance to bind.
 	 * @return True on success; otherwise false.
 	 */
-	static bool BindGlobal(uint32_t instance_id);
+	static bool BindGlobal(uint64_t instance_id);
 
 	/**
 	 * @brief Binds the instance with the given id for use. Must be done before setting
@@ -166,7 +166,7 @@ public:
 	 * @param instance_id The identifier of the instance to bind.
 	 * @return True on success; otherwise false.
 	 */
-	static bool BindInstance(uint32_t instance_id);
+	static bool BindInstance(uint64_t instance_id);
 
 	static void Destroy(const char* shader_name);
